Fix #24: replace numeric #defines in Config.h with constexpr
HEAP_UDP_PORT, FETCH_INTERVAL (now FETCH_INTERVAL_MS) and MAX_RUNTIME_TARGETS
are typed constants. SCREEN_DEBUG/MEMORY_DEBUG stay macros because they are
tested with #if/#ifdef. MAX_RUNTIME_TARGETS is size_t to match the length
constants in Structs.h and its array-bound uses. Verified with pio run.

+++ src/RuntimeBuild/ConfigLoader.cpp
@@ -142,7 +142,7 @@ ConfigResult parseBuildInfoStationMode(RuntimeConfig& outConfig, JsonObjectConst
         return failure("build_info.json missing stationId");
     }
 
-    for (int i = 0; i < MAX_RUNTIME_TARGETS; i++) {
+    for (size_t i = 0; i < MAX_RUNTIME_TARGETS; i++) {
         copyBounded(outConfig.bus.targets[i].stationId, sizeof(outConfig.bus.targets[i].stationId), stationId);
         outConfig.bus.targets[i].line[0] = '\0';
         resetArrivalState(outConfig.bus.targets[i]);
